refactored knowledge bases to a common superclass like a boss
